Add NPM distribution packages for protoc-gen-connectweb and protoc-gen-es
This is basically just a bunch of configuration files.

We compile both commands for multiple platforms, and let NPM sort out the right one to install.
This pattern is used by other Go commands in the NPM ecosystem, for example esbuild.

There is a script to compile and update all packages: make/scripts/go-build-npm.js