Rollback makeServiceType()
We introduced a mapped type for services, which allowed us to make some parts of static service definitions optional, and add the "localName" to each method.

However, this caused a few problems with mapping that type further, and lead to losing deprecation and JSDoc information.

This rolls back on the change, making generated service definitions plain const definitions again.