Add transport abstraction, interceptors, and an error type
The original plan here was to allow creating a "base client" that could be further shaped to a promise-client or callback-client, or other user-defined shapes.

Unfortunately the code read like pasta, so the magic was dialed down to only one type-mapping concept: From static service definition to a client.

Instead, we introduce "ClientTransport", so that concrete clients can be independent of the protocol.

Transports can offer interceptors, which operate on the same level as clients interact with transports.