Document ClientCallOptions and rename "abort"
"abort" is a fitting name for a property with an AbortSignal, but users will need to work with both an AbortController and an AbortSignal.

Let's not go against the grain and use "signal" as a property name. It is used in the fetch API too.

Also: Remove unused argument from overloaded method implementation.