Tighten type constraints
To correctly capture messages, we need to constrain with `T extends Message<T>`, not just `T extends T`. This is because the concrete type needs to be recursively passed down through the getType() method of message instances.

This is a bit unfortunate, but users can still use the simple constraints `T extends Message` if they do not rely on getType() (directly or indirectly). They can also rely on `Message` if they do not rely on fields.

This also renames `DynamicMessage` to `AnyMessage`. It has fewer characters and fits.

This removes the last TODO markers in the TypeScript source.