Add ConnectError.from()
Converts the argument into a ConnectError, if it is not already one.

This method is convenient for error handling in TypeScript. Connect only
raises ConnectErrors, the type system is not aware of that, forcing you
to check the type of the raised error:

```typescript
catch(throwable) {
  if (throwable instanceof ConnectError) {
    e.details;
  } else {
    // ?
  }
}
```

Example usage:

```typescript
catch(throwable) {
  const e = ConnectError.from(throwable);
  e.details;
}
```