Fix: raise a ConnectError with Code.Canceled from StreamResponse.read() instead of AbortError
Our goal is to only raise ConnectError. Due to on oversight, we do not convert an AbortError to a ConnectError in StreamResponse.read(). That means that a `for await` loop can raise an unwanted AbortError in some situations (response headers arrived, but the body is still being read).

This fixes the bug in the gRPC-web transport and in the Connect transport.