Switch tests for @bufbuild/connect-node to run on Node v16.15.0
We require the Headers class to be present. It was first added to Node v16.15.0 with the `--experimental-fetch` flag, and made available by default with Node v18.0.0.

This switches the tests to run on the older version of Node, and makes an alternative Headers implementation available in the global scope.

We have yet to decide which version of Node is going to be the oldest supported one, but so far, recommending v18 and requiring a polyfill (or the `--experimental-fetch` flag) for older versions seems reasonable.