Use web headers consistently
We convert Node headers to web headers in the protocol handlers. This removes a couple of places where we still accessed the Node headers instead of the already converted headers. This is desirable because it removes ambiguity.
This also removes a couple of places where we convert to web headers twice.