Update transform streams
This makes adds overrides to transformSerialize() and transformParse(), joining the functionality of the variants with support for end-of-stream messages.

This adds a feature: If the "endSerialization" argument is null, the transform raises an error if the end-stream flag is set.

This adds Serialization objects for gRPC-web trailers and Connect's EndStreamResponse.