Refactor message size limits and fix an enveloping bug
This moves message size closer to where they are needed: `writeMaxBytes` is honored right after serialization of a message. `readMaxBytes` is honored when reading an envelope, and when decompressing it. We look at message sizes in less places now.

This also renames the iterable functions for a more consistent API, fixes a bug when reading envelopes (the bug did not affect  @bufbuild/connect-web).