Extend universal handlers to support pre-parsed JSON requests
Fastify, express, and other servers parse `application/json` automatically. To support this behavior, this changes the "body" property of universal server requests to accept either a byte stream, or a JSON value.

The Connect protocol handler - the only one that needs to handle this behavior - was updated to support JSON request bodies.

Note that the `function createHandler()` exported by `@bufbuild/connect-node` does not make use of the new capability yet - this will be implemented in a follow-up.

Other changes:
- removes superseded `parseContentType` functions
- moves logic for protocol negotiation from the Node.js world to `createUniversalHandler()`
- adds `ContentTypeMatcher` to cache RegExp invocations
- makes I/O limits available outside of streams, for Connect unary