Switch code size benchmark to use a JS brotli implementation
This switches the code size benchmark to use the [brotli npm package](https://www.npmjs.com/package/brotli) (an Emscripten port of the [original C++ implementation](https://github.com/google/brotli)) instead of the Node.js zlib module.

We have seen slight variance of a couple of bytes between runs with the zlib module, and the JS implementation should produce stable numbers. This allows us to enable diffchecks on the results, so that CI can ensure the report is always current.