Add query parameter utilities for Connect Get
It's necessary to handle parsing and decoding URL query parameters
ourselves due to the fact that JavaScript/the DOM lack the functionality
needed to handle potentially binary data in URL-encoded form.

In the future, we may be able to improve performance by determining if
we need to do this using a regular expression.